udpated home page details
