+++ client/src/components/UpdateBtn/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+import "./style.css";
+
+function UpdateBtn(props) {
+    return (
+        <span className="update-btn" {...props} role="button" tabIndex="0">
+            √
+        </span>
+    );
+}
+
+export default UpdateBtn;
