+++ client/src/components/UpdateBtn/style.css
@@ -1,0 +1,4 @@
+.update-btn {
+    float: right;
+    color: #F2AF29;
+}
